refactor: Rename meshing_workflow, workflow and 2 meshing classes.
