+++ src/ansys/fluent/core/session/solver.py
@@ -67,7 +67,7 @@ from ansys.fluent.core.system_coupling import SystemCoupling
 from ansys.fluent.core.utils.fluent_version import (
     get_version_for_file_name,
 )
-from ansys.fluent.core.workflow import ClassicWorkflow
+from ansys.fluent.core.workflow_old import ClassicWorkflow
 
 if TYPE_CHECKING:
     from ansys.fluent.core.fluent_connection import FluentConnection
